+++ pkgs/test/test/runner/configuration/platform_test.dart
@@ -4,7 +4,6 @@
 
 @TestOn('vm')
 import 'dart:convert';
-import 'dart:io';
 
 import 'package:test/test.dart';
 import 'package:test_core/src/util/exit_codes.dart' as exit_codes;
@@ -205,10 +204,8 @@ void main() {
       var test = await runTest(['.']);
       expect(
           test.stdout,
-          containsInOrder([
-            '+0: .${Platform.pathSeparator}test_foo.dart: test_foo',
-            '+1: All tests passed!'
-          ]));
+          containsInOrder(
+              ['+0: ./test_foo.dart: test_foo', '+1: All tests passed!']));
       await test.shouldExit(0);
     });
 
@@ -243,10 +240,8 @@ void main() {
       var test = await runTest(['.']);
       expect(
           test.stdout,
-          containsInOrder([
-            '+0: .${Platform.pathSeparator}foo_test.dart: foo_test',
-            '+1: All tests passed!'
-          ]));
+          containsInOrder(
+              ['+0: ./foo_test.dart: foo_test', '+1: All tests passed!']));
       await test.shouldExit(0);
     });